alerts

+++ WypozyczalniaSamochodow/src/main/java/pl/wypozyczalnia/samochodow/modelFx/PracownicyModel.java
@@ -5,17 +5,26 @@ import javafx.beans.property.SimpleObjectProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import pl.wypozyczalnia.samochodow.database.dao.PracownicyDao;
+import pl.wypozyczalnia.samochodow.database.dao.WypozyczeniaDao;
 import pl.wypozyczalnia.samochodow.database.dbutils.DbManager;
 import pl.wypozyczalnia.samochodow.database.models.Pracownicy;
+import pl.wypozyczalnia.samochodow.database.models.Wypozyczenia;
+import pl.wypozyczalnia.samochodow.utils.Alerts;
 import pl.wypozyczalnia.samochodow.utils.converters.ConverterPracownik;
+import pl.wypozyczalnia.samochodow.utils.converters.ConverterWypozyczenia;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 public class PracownicyModel {
 
     private ObservableList<PracownicyFx> observableListPracownicy = FXCollections.observableArrayList();
     private ObjectProperty<PracownicyFx> objectPropertyPracownicy = new SimpleObjectProperty<>(new PracownicyFx());
     private ObjectProperty<PracownicyFx> objectPropertyPracownicyEdit = new SimpleObjectProperty<>();
+
+    private List<WypozyczSamochodFX>           wypozyczeniaFxList         = new ArrayList<>();
 
 
     public void init(){
@@ -52,11 +61,33 @@ public class PracownicyModel {
 
 
     public void deleteById(){
+        WypozyczeniaDao wypozyczeniaDao = new WypozyczeniaDao();
+        wypozyczeniaFxList.clear();
+        List<Wypozyczenia> wypozyczenia = wypozyczeniaDao.queryForAll(Wypozyczenia.class);
+        wypozyczenia.forEach(e -> {
+            wypozyczeniaFxList.add(ConverterWypozyczenia.convertToWypozyczeniaFx(e));
+        });
 
-        PracownicyDao pracownicyDao = new PracownicyDao();
-        pracownicyDao.deleteById(Pracownicy.class, objectPropertyPracownicy.getValue().getId());
-        init();
-        DbManager.closeConnectionSource();
+        if(filterPredicate(predicatePracownik())) {
+            PracownicyDao pracownicyDao = new PracownicyDao();
+            pracownicyDao.deleteById(Pracownicy.class, objectPropertyPracownicy.getValue().getId());
+            init();
+            DbManager.closeConnectionSource();
+        }else{
+            Alerts.warrningAlert("BŁĄD", "pracownik jest użyty w bazie");
+        }
+    }
+
+    private Predicate<WypozyczSamochodFX> predicatePracownik() {
+        return wypozyczSamochodFX -> wypozyczSamochodFX.getKlienciFx().getId() == getObjectPropertyPracownicy().getId();
+    }
+
+    private boolean filterPredicate(Predicate<WypozyczSamochodFX> predicate) {
+        List<WypozyczSamochodFX> list = wypozyczeniaFxList.stream().filter(predicate).collect(Collectors.toList());
+        if(list.isEmpty()){
+            return true;
+        }
+        return false;
     }
 
     public ObservableList<PracownicyFx> getObservableListPracownicy() {
